feat : 자산정보 기입 및 조회 구현, 마이페이지, 자산정보 기입 페이지 UI 구성

+++ src/main/java/kbits/kb04/parasol/users/dto/UsersDto.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 @Getter
 public class UsersDto {
 
-	private int user_no,user_age;
+	private Long user_no,user_age;
 	private String user_id,user_name,user_pw;
 	
 }
